requirements.txt
asgiref==3.8.1
Django==5.2.1
pillow==11.2.1
sqlparse==0.5.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.2
